Refactor WordChallenge component and related files; implement InputsOTP and useCooldown hooks for enhanced functionality
