+++ src/components/wordChallenge/wordChallenge.module.scss
@@ -33,7 +33,7 @@
         span {
             background: linear-gradient(90deg, #74ffd0, #3ac8be);
             font-weight: bold;
-            -webkit-background-clip: text;
+            background-clip: text;
             -webkit-text-fill-color: #0000;
         }
     }
@@ -48,12 +48,11 @@
         border-radius: 10px;
         overflow: hidden;
         padding: 2px;
+        transition: all 0.3s ease-in-out;
         position: relative;
 
         &::before {
             content: "";
-            background: linear-gradient(83.23deg, #feb154 -6.32%, #d263ee 23.89%, #5d5af6 67.88%, #09e0a0 107.67%),
-                linear-gradient(0deg, #0000004d, #0000004d);
             height: calc(100% + 4px);
             left: -2px;
             position: absolute;
@@ -63,7 +62,6 @@
 
         input {
             border-radius: inherit;
-            border: none;
             color: #521a93;
             font-size: 42px;
             height: 60px;
@@ -71,6 +69,9 @@
             width: 40px;
             text-align: center;
         }
+        &:has(input:focus){
+            background: linear-gradient(83.23deg, #feb154 -6.32%, #d263ee 23.89%, #5d5af6 67.88%, #09e0a0 107.67%),
+                linear-gradient(0deg, #0000004d, #0000004d);        }
     }
 }
