+++ src/services/challenges/challenges.ts
@@ -16,6 +16,12 @@ const getChallengesByUser = async (ctx) => {
   return data;
 }
 
+const getLengthWord = async (ctx) => {
+  const url = BASE_URL + '/word-length'
+  const data = await get(ctx, url);
+  return data;
+}
+
 const sendWordChallenge = async (ctx) => {
   const url = BASE_URL + '/word'
   const data = await post(ctx, url, {
@@ -25,4 +31,4 @@ const sendWordChallenge = async (ctx) => {
   return data;
 }
 
-export { getChallenges, getChallengesByUser, sendWordChallenge }
+export { getChallenges, getChallengesByUser, sendWordChallenge, getLengthWord }
